feat: support diagram (mermaid)

+++ packages/vue/src/extensions/char-command/slash/view.vue
@@ -49,6 +49,7 @@ import useI18n from '../../../hooks/useI18n'
 import useTiptap from '../../../hooks/useTiptap'
 import { ODivider, OIcon, OList, OListItem } from '../../../components/index'
 import { BasicBlocks } from '../../../constants/block'
+import { EmptyDiagram } from '../../../constants/empty-block'
 
 export default {
   props: {
@@ -107,35 +108,21 @@ export default {
     runCommand(item: Indexable) {
       const focus = this.editor.chain().focus().deleteRange(this.range)
       const commands = this.editor.commands
-      switch (item.value) {
-        case 'blockMath':
-          commands.deleteRange(this.range)
-          this.editor.commands.insertBlockMath({ latex: '' })
-          break
-        case 'codeBlock':
-          commands.deleteRange(this.range)
-          this.editor.commands.setCodeBlock({ language: 'bash' })
-          break
-        case 'content':
-          commands.deleteRange(this.range)
-          this.editor.commands.insertContent(item.options.content)
-          break
-        case 'emoji':
-          commands.deleteRange(this.range)
-          this.editor.commands.insertContent(':')
-          break
-        case 'inlineMath':
-          commands.deleteRange(this.range)
-          this.run(this.editor, 'inlineMath')
-          // this.onInlineMath()
-          break
-        case 'taskList':
-          commands.deleteRange(this.range)
-          this.editor.commands.toggleTaskList()
-          break
-        default:
-          this.onCommand(commands, focus, item.value, item.options)
-          break
+      const blocks = [
+        'blockMath',
+        'codeBlock',
+        'content',
+        'details',
+        'diagram',
+        'emoji',
+        'inlineMath',
+        'taskList',
+      ]
+      if (blocks.includes(item.value)) {
+        commands.deleteRange(this.range)
+        this.run(this.editor, item.value, item.options)
+      } else {
+        this.onCommand(commands, focus, item.value, item.options)
       }
     },
     onKeyDown({ event }) {
